Refactor hustle market slice into modular modules
